refactor(tracing): replace TracingStatus with SpanStatus and introduce TransactionSpanWrapper
- Removed the TracingStatus enum and its related functionality, replacing it with SpanStatus for better consistency in span status handling.
- Introduced TransactionSpanWrapper to manage transaction lifecycles, allowing for nested transactions and improved span management.
- Updated LegacySpanWrapper to utilize the new SpanStatus and TransactionSpanWrapper, enhancing the tracing capabilities.
- Adjusted exports in tracing.dart to include the new transaction span wrapper and removed the legacy tracing status.
- Updated tests to reflect changes in span status handling and transaction management.
