+++ packages/drift/lib/src/utils/constants.dart
@@ -1,2 +1,1 @@
 const String integrationName = 'sentryDriftTracing';
-const String loggerName = 'sentry_drift';
