+++ packages/drift/lib/src/utils/internal_logger.dart
@@ -1,0 +1,6 @@
+import 'package:meta/meta.dart';
+import 'package:sentry/sentry.dart';
+
+/// Logger for the Sentry Drift integration.
+@internal
+const internalLogger = SentryInternalLogger('sentry_drift');
